+++ react/components/Footer/CHANGELOG.md
