+++ react/components/Footer/README.md
