+++ react/components/Footer/components/FooterBadgeList.js
