+++ react/components/Footer/components/FooterPaymentFormList.js
