+++ react/components/Footer/index.js
