+++ react/components/Footer/propTypes.js
